Remove gulp, use npm commands instead
Gulk has lost its value for me; requires a lot more packages to be installed while it doesn't bring much in the way of benefit, and actually makes some things more difficult.
