+++ .babelrc
@@ -2,7 +2,11 @@
   "presets": [
     "@babel/preset-env"
   ],
-  "plugins": [
-    "istanbul"
-  ]
+  "env": {
+    "test": {
+      "plugins": [
+        "istanbul"
+      ]
+    }
+  }
 }
