+++ package.json
@@ -12,10 +12,12 @@
     "dist/three-sprite-texture-atlas-manager.umd.min.js.map"
   ],
   "scripts": {
-    "build": "gulp build",
-    "test": "mocha",
-    "coverage": "nyc mocha",
-    "watch": "gulp watch"
+    "start": "cross-env NODE_ENV=test mocha",
+    "test": "cross-env NODE_ENV=test mocha",
+    "watch": "nodemon --ext js --watch ./ --exec 'cross-env NODE_ENV=test mocha || true' --delay 1",
+    "build": "rollup -c",
+    "docs": "jsdoc2md --no-cache --template api.hbs --files 'src/**/*.js' > docs/API.md",
+    "coverage": "cross-env NODE_ENV=test nyc mocha"
   },
   "repository": {
     "type": "git",
@@ -48,16 +50,6 @@
     "chai": "^4.3.3",
     "cross-env": "^7.0.3",
     "del": "^6.0.0",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
-    "gulp-debug": "^4.0.0",
-    "gulp-eslint": "^6.0.0",
-    "gulp-istanbul": "^1.1.3",
-    "gulp-mocha": "^8.0.0",
-    "gulp-plumber": "^1.2.1",
-    "gulp-rename": "^2.0.0",
-    "gulp-shell": "^0.8.0",
-    "gulp-sourcemaps": "^3.0.0",
     "jsdoc-to-markdown": "^7.0.0",
     "mocha": "^8.3.2",
     "nodemon": "^2.0.7",
